+++ src/test/java/ro/unibuc/hello/controller/unit/TaskControllerTest.java
@@ -1,4 +1,4 @@
-package ro.unibuc.hello.controller;
+package ro.unibuc.hello.controller.unit;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.builder.EqualsBuilder;
@@ -9,6 +9,7 @@ import org.mockito.*;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
+import ro.unibuc.hello.controller.TaskController;
 import ro.unibuc.hello.data.*;
 import ro.unibuc.hello.dto.CreateTask;
 import ro.unibuc.hello.dto.TaskDto;
@@ -28,7 +29,7 @@ import static org.mockito.Mockito.when;
 class TaskControllerTest {
 
 	@InjectMocks
-	TaskController taskController;
+    TaskController taskController;
 
 	@Mock
 	TaskRepository taskRepository;
