+++ src/test/java/ro/unibuc/hello/controller/unit/UserControllerTest.java
@@ -1,4 +1,4 @@
-package ro.unibuc.hello.controller;
+package ro.unibuc.hello.controller.unit;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -7,6 +7,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
+import ro.unibuc.hello.controller.UserController;
 import ro.unibuc.hello.data.User;
 import ro.unibuc.hello.data.UserRepository;
 import ro.unibuc.hello.dto.UserProfile;
